requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
numpy==1.24.4
lime==0.2.0.1
scikit-learn==1.3.2
matplotlib==3.8.2
pyyaml==6.0.1
opencv-python==4.9.0.80
Pillow==10.1.0
seaborn==0.13.0 